feat-38 : 사용성 테스트 (request hook)
