+++ src/models/ITokenInfo.ts
@@ -1,0 +1,4 @@
+export interface ITokenInfo {
+  accessToken: string;
+  refreshToken: string;
+}
